requirements.txt

langchain>=0.1.16
openai>=1.14.3
anthropic>=0.25.5
google-generativeai>=0.4.1
tiktoken>=0.6.0
sentence-transformers>=2.6.1
faiss-cpu>=1.7.4
numpy>=1.24
gradio>=4.25
